Let the garbage collector handle cleanup
The current behavior here is to create a curl session, call a function,
then clean up the session immediately after the function returns.
Unfortunately, if the function we pass to this returns a Lwt promise
(and probably whatever Async does as well) then the session no longer
exists when the system tries to resolve the promise, resulting in an
"invalid argument" exception from curl.  I initially tested this by just
removing the cleanup call, but this is better and doesn't leak memory.